moved all of the logic that was initally on the Home page to the MainContent component due to a production bug where the ip address is fetch from the aws server and not the client-side browser; this was done by moving code and some refactoring,and installing the useSWR library for client-side fetch logic
