+++ src/api/fetchWeatherData.js
@@ -1,7 +1,7 @@
 const fetchWeatherData = async(passIpAddress) => {
     try {
-      const weatherURL = 'https://api.weatherapi.com/v1/'
-      const fetchForecast = await fetch(`${weatherURL}forecast.json?key=${process.env.WEATHER_API_KEY}&q=${passIpAddress}&days=1`)
+      const weatherURL = `${process.env.NEXT_PUBLIC_WEATHER_API_BASE_URL}`
+      const fetchForecast = await fetch(`${weatherURL}?key=${process.env.WEATHER_API_KEY}&q=${passIpAddress}&days=1`)
       const data = await fetchForecast.json()
       return data
     } catch (error) {
